Move geometry to a individual folder
