+++ svgdatagrabber/geometry/__init__.py
@@ -1,0 +1,3 @@
+from .geometrybase import GeometryBase
+from .line import Line, Segment, Ray, ExtendedSegment, ExtendedRay
+from .point import Point, Vector
